+++ store/service/src/Controllers/User.php
@@ -4,6 +4,8 @@ namespace Controllers;
 use PDO;
 use Exception;
 
+use Utils\Authenticate;
+
 class User extends \Utils\BaseController {
   protected function createTable() {
     try {
@@ -21,11 +23,17 @@ class User extends \Utils\BaseController {
       $sth->execute();
     }
     catch (Exception $e) {
-      error_log ('User::createTable: ' . $e->getMessage());
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 
   public function getAll() {
+    if (false === Authenticate::isAuth()) {
+      //TODO: Use a Response class
+      header('HTTP/1.0 401 Unauthorized');
+      return array('errorCode'=> 401, 'errorMessage' => 'Not Authorized');
+    }
     try {
       $sth = $this->pdo->prepare('
         SELECT
@@ -38,11 +46,17 @@ class User extends \Utils\BaseController {
       return $sth->fetchAll(PDO::FETCH_ASSOC);
     }
     catch (Exception $e) {
-      error_log ('User::getAll: ' . $e->getMessage());
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 
   public function get($id) {
+    if (false === Authenticate::isAuth()) {
+      //TODO: Use a Response class
+      header('HTTP/1.0 401 Unauthorized');
+      return array('errorCode'=> 401, 'errorMessage' => 'Not Authorized');
+    }
     try {
       $sth = $this->pdo->prepare('
         SELECT
@@ -53,16 +67,22 @@ class User extends \Utils\BaseController {
         WHERE id LIKE :id;
       ');
       $sth->execute(array('id' => $id));
-      return $sth->fetch(PDO::FETCH_ASSOC);
+      $resource = $sth->fetch(PDO::FETCH_ASSOC);
+      if (false === $resource) {
+        //TODO: Return a Response object
+        header('HTTP/1.0 404 Not Found');
+        return array('errorCode' => 404, 'errorMessage' => 'Not Found');
+      }
+      return $resource;
     }
     catch (Exception $e) {
-      error_log ('User::get: ' . $e->getMessage());
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 
   // Create
   public function put($id, $user) {
-    //FIXME: Check user permissions
     try {
       $sth = $this->pdo->prepare('
         INSERT INTO `users`(
@@ -85,16 +105,21 @@ class User extends \Utils\BaseController {
         'update_date'  => $user->update_date,
       ));
       //FIXME: Return 201 or 204
+      header('HTTP/1.0 201 Created');
     }
     catch (Exception $e) {
-      error_log ('User::put: ' . $e->getMessage());
-      //TODO: Return 400?
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 
   // Update
   public function post($id, $user) {
-    //FIXME: Check user permissions
+    if ($id === Authenticate::isAuth()) {
+      //TODO: Use a Response class
+      header('HTTP/1.0 401 Unauthorized');
+      return array('errorCode'=> 401, 'errorMessage' => 'Not Authorized');
+    }
     try {
       $sth = $this->pdo->prepare('
         INSERT OR REPLACE INTO `users`(
@@ -116,15 +141,20 @@ class User extends \Utils\BaseController {
         'create_date'  => $user->create_date,
         'update_date'  => $user->update_date,
       ));
+      header('HTTP/1.0 204 No Content');
     }
     catch (Exception $e) {
-      error_log ('User::post: ' . $e->getMessage());
-      //TODO: Return 400?
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 
   public function delete($id) {
-    //FIXME: Check user permissions
+    if ($id !== Authenticate::isAuth()) {
+      //TODO: Use a Response class
+      header('HTTP/1.0 401 Unauthorized');
+      return array('errorCode'=> 401, 'errorMessage' => 'Not Authorized');
+    }
     try {
       $sth = $this->pdo->prepare('
         DELETE FROM `users` WHERE `id` LIKE :id
@@ -135,7 +165,8 @@ class User extends \Utils\BaseController {
       //FIXME: Return 204
     }
     catch (Exception $e) {
-      error_log ('User::delete: ' . $e->getMessage());
+      error_log (__METHOD__ . ': ' . $e->getMessage());
+      header('HTTP/1.0 503 Service Unavailable');
     }
   }
 }
